# docs/swagger.yaml
basePath: /
definitions:
  models.Priority:
    enum:
    - low
    - medium
    - high
    type: string
    x-enum-varnames:
    - PriorityLow
    - PriorityMedium
    - PriorityHigh
  models.Status:
    enum:
    - pending
    - in_progress
    - completed
    - cancelled
    type: string
    x-enum-varnames:
    - StatusPending
    - StatusInProgress
    - StatusCompleted
    - StatusCancelled
  models.Task:
    properties:
      created_at:
        type: string
      description:
        type: string
      due_date:
        type: string
      id:
        type: string
      priority:
        $ref: '#/definitions/models.Priority'
      status:
        $ref: '#/definitions/models.Status'
      title:
        type: string
      updated_at:
        type: string
    type: object
  services.CreateTaskInput:
    properties:
      description:
        type: string
      due_date:
        type: string
      priority:
        type: string
      title:
        type: string
    type: object
  services.UpdateTaskInput:
    properties:
      description:
        type: string
      due_date:
        type: string
      priority:
        type: string
      status:
        type: string
      title:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API RESTful de gerenciamento de tarefas desenvolvida em Go
  title: Desafio To Do List API
  version: "1.0"
paths:
  /tasks:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Task'
            type: array
      summary: Listar tarefas
      tags:
      - tasks
    post:
      consumes:
      - application/json
      description: Cria uma nova tarefa
      parameters:
      - description: Dados da tarefa
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/services.CreateTaskInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Task'
      summary: Criar tarefa
      tags:
      - tasks
  /tasks/{id}:
    delete:
      parameters:
      - description: ID da tarefa
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      summary: Deletar tarefa
      tags:
      - tasks
    get:
      parameters:
      - description: ID da tarefa
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Task'
      summary: Buscar tarefa por ID
      tags:
      - tasks
    put:
      consumes:
      - application/json
      parameters:
      - description: ID da tarefa
        in: path
        name: id
        required: true
        type: string
      - description: Dados da tarefa
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/services.UpdateTaskInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Atualizar tarefa
      tags:
      - tasks
swagger: "2.0"